Initial commit of `yesiscan`
This is a new project related to license scanning. It was authored by
James Shubin (https://purpleidea.com/) while at Amazon.ca and is being
released in the hope that it will be useful for others.

It is still early in the project, but watch this space, and get involved
if this sort of thing interests you!