backend: Add scancode backend
This adds an imperfect, but functional scancode backend. It runs things
properly, but we could improve performance significantly by
parallelizing the file iteration and perhaps by running scancode as an
api server that receives a stream of requests. A directory traversal
interface could also be another way to write this.