main, lib: Add fancy profiles based filtering
This adds some fancy filtering based on a new concept called profiles.
It's not clear if there is a better name for this, so we will keep it as
such for now. The README is fairly complete in how this all works.

Additionally we extend the licensing library to support new matching.