web: Add a webui for running scans
This monster patch adds a web frontend to the project. It can be used as
a standalone alternative for users who prefer this over a cli. It can
eventually be extended to run as a hosted service for users who don't
want to run it locally.

This also adds the summary display to the outputs. This allows the
observer to see a quick recap of all the different licenses found.

This also includes some other patches that were mixed in at the same
time.