lib, web, cmd: Allow output to an S3 bucket
This patch allows you to write reports directly to an S3 bucket. You
are solely responsible for the privacy implications should you use this
option with private data.

In addition, this does a major rework of the flags passing, so that it
is unified across the project. Now the backends will be more consistent.