lib, backend: Plumb through the concept of skipped/passed files
This lets us track which files were skipped due to simply being void of
found data and which caused a scanning error. The latter is important
because we'd like to prevent a single file issue from nuking our entire
scan. Instead, we collect the specific errors and notify the users at
the end. This type of thing may include a single file with an absurdly
long line (that we don't wish to read into memory to scan) or a file
which is pretending to be a zip file but which is actually not.

This also refactors some of the output template profiles at the same
time.