cmd: yesiscan: Detect whether config was just changed
When the config was just changed, even though it will probably not be
expired yet, we still want to keep downloading from stage 2 in the
config downloader because there might be new files there. Since we are
just looking at the core config to determine if we're expired, we should
make sure to run for the child files right after this happens.