cmd: yesiscan: Allow auto-update of binary
This builds an auto-update mechanism for the yesiscan binary. The new
versions need to be listed in the config file, along with their URL's,
and you have to explicitly enable this for it to work.

If successful, this will exec the new version when it is found.

This has not been tested on Windows or Mac.