lib, iterator: Add iterator error and support in zip
This adds a iterator error functionality to let an iterator return a
non-fatal error which gets propagated upwards. Unfortunately there are
circumstances which make it impossible for an iterator to fail, but for
which we still want to complete a scan with partial results. The common
case is one in which we have a corrupt or invalid zip file for example.
An empty zip file (zero bytes) is an example of this. When this happens,
we collect all the iterator errors and propagate them to the user to see
in the report.